+++ spec/behavior/emitter.js
@@ -1,17 +1,28 @@
+var _ = require( 'lodash' );
+
 module.exports = function emitter( name ) {
 
 	var handlers = {};
 
 	function raise( ev ) {
 		if ( handlers[ ev ] ) {
-			handlers[ ev ].apply( undefined, Array.prototype.slice.call( arguments, 1 ) );
+			var args = Array.prototype.slice.call( arguments, 1 );
+			_.each( handlers[ ev ], function( handler ) {
+				if ( handler ) {
+					handler.apply( undefined, args );
+				}
+			} );
 		}
 	}
 
 	function on( ev, handle ) {
-		handlers[ ev ] = handle;
-		return { unsubscribe: function() {
-				delete handlers[ ev ];
+		if ( handlers[ ev ] ) {
+			handlers[ ev ].push( handle );
+		} else {
+			handlers[ ev ] = [ handle ];
+		}
+		return { unsubscribe: function( h ) {
+				var r = handlers[ ev ].splice( _.indexOf( handlers[ ev ], h || handle ) );
 			} };
 	}